refactor: 窓の可視性・位置・サイズ・wake を window_coordinator へ集約する (#749 C2)
mod.rs から 8 関数（show/hide・save_placement_relative・wake_main/wake_results・
position_results_below_main・results_available_height・read_metrics）、main.rs から
position_on_target_monitor、view.rs から drive_results_window / max_results を
window_coordinator.rs へ移す。挙動は変えない。

同伴させるヘルパーの規則を 1 本にした（例外なし）: 移設する関数がその中でしか使わない
ものは一緒に運び、複数モジュールから消費されるものは残す。ゆえに read_metrics と
max_results は運び、read_visual と window_width は残る。listener の登録は 4 関数とも
mod.rs に残す——register_hide_listener だけを動かすと、show_egui_main と wake_main を
呼ぶ register_initial_hotkey_failure_listener が残る理由を説明できなくなる。

z-order は集約しない（main は commands/window.rs が set_always_on_top を直接叩き、
results は ResultsWindow が持つ・層が違う）。main 窓のサイズ適用も view に残す
（意図的な 2 導出・ADR-0007 却下 1）。どちらも新モジュールの //! に明記した。

drive_results_window は自由関数化し、読み点の制約を持つ 2 値（result_count は
クリック逆流の消費後・plain_hidden は消費前）と幾何 2 値だけを引数に取る。
max_results は制約を持たないため内側で読む——引数を増やすほど、呼び出し側で
読み点の違う値を並べて書きたくなる。

移設で落ちると永久に気づかない #[cfg(not(windows))] の双子 arm は、移設前後で
合計 7 件のまま（CI は Windows のみ）。cargo test -p snotra = 176 passed。
